pyproject.toml
[tool.poetry]
name = "api-for-color-palettes"
version = "0.1.0"
description = ""
authors = ["vladimirmac <vladimirbodrin89@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
djangorestframework = "^3.15.1"
python-dotenv = "^1.0.1"
gunicorn = "^21.2.0"
drf-spectacular = "^0.27.2"
psycopg2-binary = "^2.9.9"
djangorestframework-simplejwt = "^5.3.1"
celery = {extras = ["redis"], version = "^5.3.6"}


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.0"
ruff = "^0.3.5"
django-debug-toolbar = "^4.3.0"
pytest-cov = "^5.0.0"
pytest-factoryboy = "^2.7.0"
pytest-django = "^4.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
